split coupon methods into their own file
